change the data type of constan_values of Pad from int into float

+++ include/caffe/layers/pad_layer.hpp
@@ -37,7 +37,7 @@ class PadLayer : public Layer<Dtype> {
   }
 
   vector<int> paddings_;
-  int constant_values_;
+  float constant_values_;
 };
 
 }  // namespace caffe
